+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ### Changelog
 
-#### 0.3.2 (Next)
+#### 0.4.0 (Next)
 
+* The `reset` and `resetting` callbacks have been removed - [@dblock](https://github.com/dblock).
+* Callbacks receive `team` and `error`, but no longer `server` - [@dblock](https://github.com/dblock).
+* [#23](https://github.com/dblock/slack-ruby-bot-server/issues/23): Fix: server-side removal of bot doesn't deactivate it - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 #### 0.3.1 (7/10/2016)